test for btnFunction first

+++ perl/openprint/administrator_project_types.pm
@@ -354,15 +354,17 @@ sub _paper_recommendations {
 
 sub categories {
 	my $ProjectTypeCategory = new openprint::ProjectTypeCategory( $param{category_id} );
-	if ( $param{btnFunction} eq 'Save' ) {
-		$variable{error} .= $ProjectTypeCategory->save(\%param);
-		foreach my $pt_id ( ref $param{projecttype_id} eq 'ARRAY' ? @{$param{projecttype_id}} : $param{projecttype_id} ) {
-			my $ProjectType = new openprint::ProjectType( $pt_id );
-			$variable{error} .= $ProjectType->save({category_id=>$ProjectTypeCategory->id()});
-		} # end foreach pt_id
-	} elsif ( $param{btnFunction} eq 'Delete' ) {
-		$variable{error} .= $ProjectTypeCategory->delete();
-	} # end if
+  if ($param{btnFunction}) {
+    if ( $param{btnFunction} eq 'Save' ) {
+      $variable{error} .= $ProjectTypeCategory->save(\%param);
+      foreach my $pt_id ( ref $param{projecttype_id} eq 'ARRAY' ? @{$param{projecttype_id}} : $param{projecttype_id} ) {
+        my $ProjectType = new openprint::ProjectType( $pt_id );
+        $variable{error} .= $ProjectType->save({category_id=>$ProjectTypeCategory->id()});
+      } # end foreach pt_id
+    } elsif ( $param{btnFunction} eq 'Delete' ) {
+      $variable{error} .= $ProjectTypeCategory->delete();
+    } # end if
+  } # end if
 	$variable{ProjectTypeCategory} = $ProjectTypeCategory;
 } # end sub categories
